fix(ci): split oversized planning_ops.py to satisfy line-limit check
Move post-validation conformance helpers into a new planning_conformance.py
module so planning_ops.py stays under the 1500-line maintainability limit.
Update executor mixin delegations and adjust two tests that patched module-level
repo/deposit names to target the new module.
